refactor: wire Profile onboarding/management variants into export files and remove barrel index
