--- src/components/Employee/Profile/index.ts
@@ -1,1 +1,0 @@
-export * from './onboarding/Profile'
