+++ src/components/Employee/SelfOnboardingFlow/SelfOnboardingComponents.tsx
@@ -4,7 +4,7 @@ import type { FlowContextInterface } from '@/components/Flow/useFlow'
 import { useFlow } from '@/components/Flow/useFlow'
 import type { BaseComponentInterface } from '@/components/Base'
 import { Landing as LandingComponent } from '@/components/Employee/Landing'
-import { Profile as ProfileComponent } from '@/components/Employee/Profile'
+import { Profile as ProfileComponent } from '@/components/Employee/Profile/onboarding/Profile'
 import { FederalTaxes as FederalTaxesComponent } from '@/components/Employee/FederalTaxes/onboarding/FederalTaxes'
 import { StateTaxes as StateTaxesComponent } from '@/components/Employee/StateTaxes'
 import { PaymentMethod as PaymentMethodComponent } from '@/components/Employee/PaymentMethod'
