+++ src/components/Employee/exports/employeeManagement.ts
@@ -4,6 +4,7 @@ export { DashboardFlow } from '../Dashboard'
 export { WorkAddress } from '../WorkAddress/management/WorkAddress'
 export type { WorkAddressProps } from '../WorkAddress/management/WorkAddress'
 export { FederalTaxes, type FederalTaxesProps } from '../FederalTaxes/management/FederalTaxes'
+export { Profile, type ProfileProps } from '../Profile/management/Profile'
 export { TerminateEmployee } from '../Terminations/TerminateEmployee/TerminateEmployee'
 export { TerminationSummary } from '../Terminations/TerminationSummary/TerminationSummary'
 export { TerminationFlow } from '../Terminations/TerminationFlow/TerminationFlow'
